fix: déplacer le dataset des catégories dans un fichier séparé pour une meilleure organisation
