+++ test/e2e/categories/categories.dataset.e2e.ts
@@ -1,0 +1,23 @@
+// 🌟 LE DATASET UNIFIÉ
+export const CategoriesDataset = {
+  user: {
+    email: "categories.e2e@example.com",
+    password: "Password123!",
+    firstName: "Cat",
+    lastName: "Tester",
+  },
+  budget: {
+    label: "Budget E2E",
+    color: "#3498db",
+  },
+  existingCategory: {
+    label: "Catégorie Existante",
+    color: "#e74c3c",
+    limitAmount: 500.0,
+  },
+  newCategory: {
+    label: "Catégorie Succès",
+    color: "#FF5733",
+    limitAmount: 100,
+  },
+};
