+++ test/e2e/categories/categories.test.context.e2e.ts
@@ -1,28 +1,6 @@
 import { PrismaService } from "src/prisma/prisma.service";
 
-// 🌟 LE DATASET UNIFIÉ
-export const CategoriesDataset = {
-  user: {
-    email: "categories.e2e@example.com",
-    password: "Password123!",
-    firstName: "Cat",
-    lastName: "Tester",
-  },
-  budget: {
-    label: "Budget E2E",
-    color: "#3498db",
-  },
-  existingCategory: {
-    label: "Catégorie Existante",
-    color: "#e74c3c",
-    limitAmount: 500.0,
-  },
-  newCategory: {
-    label: "Catégorie Succès",
-    color: "#FF5733",
-    limitAmount: 100,
-  },
-};
+import { CategoriesDataset } from "./categories.dataset.e2e";
 
 export class CategoriesTestContext {
   public userId: string;
@@ -32,7 +10,7 @@ export class CategoriesTestContext {
   constructor(private readonly prisma: PrismaService) {}
 
   async init(): Promise<void> {
-    await this.cleanup(); // 👈 Nettoyage préventif crucial
+    await this.cleanup(); // Nettoyage préventif crucial
 
     const bcrypt = await import("bcrypt");
     const hashedPassword = await bcrypt.default.hash(CategoriesDataset.user.password, 10);
@@ -73,8 +51,6 @@ export class CategoriesTestContext {
     });
 
     if (user) {
-      // Suppression sécurisée en cascade
-      await this.prisma.transaction.deleteMany({ where: { userId: user.id } });
       await this.prisma.category.deleteMany({ where: { userId: user.id } });
       await this.prisma.budget.deleteMany({ where: { userId: user.id } });
       await this.prisma.user.delete({ where: { id: user.id } });
